requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
confluent-kafka>=2.3.0
pydantic>=2.5.0
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.0
librosa>=0.10.0
numpy>=1.24.0
